% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_evidence_max.R
\name{bm_evidence_max}
\alias{bm_evidence_max}
\title{Bayesian Analyses of Simple and Balanced Linear Mixed-Effects Models.}
\usage{
bm_evidence_max(mu1,mu2,beta0,n,w,Q1,Q2,Mn)
}
\arguments{
\item{a}{bla bla}
}
\value{
bla bla
}
\description{
Bayesian analysis of simple linear mixed-effects model (LMM) applied
to balanced data, via a closed form posterior distribution.
}
\details{
See Gangsei & Vinje ... for details.
}
\note{
bla bla, use function nlm()
}
\examples{
\dontrun{

}



}
\references{
Gangsei, L.E.& Vinje, H. 2025. A closed form solution for
Bayesian analysis of a simple linear mixed model.
}
\author{
Hilde Vinje\cr
hilde.vinje@nmbu.no\cr
+47 959 01 450

Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
