% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlmer.R
\name{bmlmer}
\alias{bmlmer}
\title{Bayesian Analyses of Simple and Balanced Linear Mixed-Effects Models.}
\usage{
bmlmer(mod_formula, data, nu1,mu1,nu2,mu2,beta0,nu0,Upsilon0)
}
\arguments{
\item{mod_form}{a two-sided linear formula object describing both the
fixed-effects and random-effects part of the model,
with the response on the left of a ~ operator and the terms,
separated by + operators, on the right.
Random-effects terms are distinguished by vertical bars (|).
So far only implemented for one grouping variable in a deign that
has to be balanced.}

\item{data}{a data frame containing the variables named in mod_form.}

\item{nu1, mu1, nu2, mu2, beta0, nu3, Upsilon0.}{Prior hyper parameters, nu1, nu2 and
nu3 are scalars with prior sample sizes for the beta, gamma and multivariate
normal distributions respectively. mu1 (0<mu1<1) and mu2 are prior expectations
for the beta and gamma distributions respectively. beta0 (length p) is prior
expected value for fixed regression parameters and Upsilon0 (p x p)
is prior variance for regression parameters up to proportionality.}

\item{nsim}{number of independent simulations in the posterior distribution.}

\item{alpha}{significance level for credibility intervals that are returned.}

\item{simreturn}{logical. If the full simulations is to be returned}

\item{empirical_bayes}{logical. If prior hyper-parameters is to be set via
empirical Bayes strategy.}

\item{nu_start, nu_max, nu_min}{optional. All vectors of length 3 setting start,
minimum and maximum numbers for the three prior sizes (nu1, nu2, and nu3)
if empirical Bayes is applied to set hyperparameters.}
}
\value{
The function returns a list with elements:\cr
\itemize{
\item Mean: vector with posterior means for parameters\cr
\item Quantiles: matrix with posterior quantiles for all parameters at
levels alpha/2, 50\% (median) and 1-alpha/2, i.e. credibility intervals
with median.\cr
\item Covariance posterior covariance matrice for the parameters.\cr
\item Correlation posterior covariance matrice for the parameters.\cr
\item simulations (only returned if simreturn == TRUE), all simulated values.\cr
\item eb_nu (only returned if empirical_bayes == TRUE), empirical Bayes results
for prior means.\cr
}
}
\description{
Bayesian analysis of simple linear mixed-effects model (LMM) applied
to balanced data, via a closed form posterior distribution.
}
\details{
See Gangsei & Vinje (2026) for details.
}
\note{
So far code and method is restricted to the simple, balanced design of
the mixed model. Hopefully it will be possible to find a solution where the
method is applicable for more complex mixed models.
}
\examples{
\dontrun{
## Clean workspace
rm(list = ls())  

## Load the bmmix packages
#devtools::install_git(url = 'https://github.com/hildvi/bmmix')

require(bmmix)

## Set dimensions and parameters in model

nrep <- 1000 # Number of simulations

### Dimentions
w <- 4
p <- 3
n <- 20

### Set parameters
sigma2 <- 4
sigma2u <- 0.5

delta <- w*sigma2u/(sigma2 + w*sigma2u)
beta <- c(0.2,2,-0.5)

### Put parameters in one vector
par_vec <- c(delta,sigma2,sigma2u,as.vector(beta))
Par_mat <- matrix(par_vec,nrep,length(par_vec),byrow = TRUE)

## Set up model and environment for evaluation
bmlmerform <- formula(y ~ X + (1|id))
ID <- matrix(rep(1:n,w),n,w,byrow=FALSE)

bayes_eval <- freq_eval <- matrix(NA,0,6, dimnames = 
                                    list(NULL,
                      c('delta','sigma2','sigma2u','(intercept)','x1','x2')))



## Start simulation 
### Set seeds to replicate results.
set.seed(15)
new_seeds <- sample(1:(10*nrep),nrep)

for(i in 1:nrep)
{
  set.seed(new_seeds[i]) #New seed 
  
  ### Simulate new design matrices and random effects
  X <- cbind(1,matrix(rnorm(n*(p-1)),n,p-1))
  U <- matrix(rep(rnorm(n,0,sqrt(sigma2u)),w),n,w,byrow=FALSE)
  E <- matrix(rnorm(n*w,0,sqrt(sigma2)),n,w)
  Y <- matrix(X\%*\%beta,n,w,byrow=FALSE) + U + E
  
  ### Put data in to data frame
  mixdata <- data.frame(y = as.vector(Y),
                        X = I(kronecker(as.matrix(rep(1,w)),X[,-1])),
                       id = as.vector(ID))
 
  ### Get the result from Bayesian inference/ empirical Bayes
  bmlmer_res <- bmlmer(mod_form = bmlmerform,data = mixdata,
                       nsim = 5000,beta0=NULL,
                       simreturn = FALSE,empirical_bayes = TRUE,
                       nu_min = c(2,0,0),nu_max = c(2.001,5,5))
  
  #print(bmlmer_res$eb_nu)
  bayes_eval <- rbind(bayes_eval,
                      rbind(bmlmer_res$Quantiles[c(1,3),],
                            bmlmer_res$Mean)) #Store results
  
  ### Get results from standard frequentist approach
  freq_mod <- lme4::lmer(formula = bmlmerform,data = mixdata)
  lme4_res <- confint(freq_mod)
  lme4_res[1:2,] <- lme4_res[1:2,]^2 
  
  if(rownames(lme4_res)[1]==".sig01"){lme4_res[1:2,] <- lme4_res[c(2,1),]}
  
  lme4_res <- t(rbind(NA,cbind(lme4_res,
                               c(as.data.frame(lme4::VarCorr(freq_mod))$vcov[c(2,1)],
                                 summary(freq_mod)$coef[,1]))))
  
  freq_eval <- rbind(freq_eval,lme4_res)# store results
}

## Table for manuscript
Tab1 <- data.frame(
  Real_value = par_vec,
  Bayes_OK_per = colSums((bayes_eval[seq(1,3*nrep,by = 3),-7]<Par_mat)&
                           (bayes_eval[seq(2,3*nrep,by = 3),-7]>Par_mat))/nrep,
  Bayes_CI_widh = apply(bayes_eval[seq(2,3*nrep,by = 3),-7]-
                          bayes_eval[seq(1,3*nrep,by = 3),-7],2,mean),
  Bayes_MSE = apply(bayes_eval[seq(3,3*nrep,by = 3),-7]-
                      Par_mat,2,var),
  Bayes_bias = apply(bayes_eval[seq(3,3*nrep,by = 3),-7]-
                       Par_mat,2,mean),
  Freq_OK_per =colSums((freq_eval[seq(1,3*nrep,by = 3),-7]<Par_mat)&
                         (freq_eval[seq(2,3*nrep,by = 3),-7]>Par_mat))/nrep,
  Bayes_CI_widh = apply(freq_eval[seq(2,3*nrep,by = 3),-7]-
                          freq_eval[seq(1,3*nrep,by = 3),-7],2,mean),
  Freq_MSE = apply(freq_eval[seq(3,3*nrep,by = 3),-7]-
                     Par_mat,2,var),
  Freq_bias = apply(freq_eval[seq(3,3*nrep,by = 3),-7]-
                      Par_mat,2,mean))


### Write table in Latex format
xtable::xtable(Tab1)

}



}
\references{
Gangsei, L.E.& Vinje, H. 2026. A closed form solution for
Bayesian analysis of a simple linear mixed model.
}
\author{
Hilde Vinje\cr
hilde.vinje@nmbu.no\cr

Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
}
