% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlmer.R
\name{bmlmer}
\alias{bmlmer}
\title{Bayesian Analyses of Simple and Balanced Linear Mixed-Effects Models.}
\usage{
bmlmer(formula, data, nu1,mu1,nu2,mu2,beta0,nu0,Upsilon0)
}
\arguments{
\item{a}{bla bla}
}
\value{
bla bla
}
\description{
Bayesian analysis of simple linear mixed-effects model (LMM) applied
to balanced data, via a closed form posterior distribution.
}
\details{
See Gangsei & Vinje ... for details.
}
\note{
bla bla
}
\examples{
\dontrun{

## Simulate data
set.seed(11)
nrep <- 10000
new_seeds <- sample(1:(10*nrep),nrep)
sigma2 <- 1.5
sigma2u <- 0.75
beta <- as.matrix(rnorm(p))
par_vec <- c(sigma2,sigma2u,as.vector(beta))
w <- 4
p <- 3
n <- 15
bmlmerform <- formula(y ~ X + (1|id))
ID <- matrix(rep(1:n,w),n,w,byrow=FALSE)

eval <- matrix(0,2,5, dimnames = 
          list(c('bmlmer','lme4'),
          c('sigma2','sigma2u','(intercept)','x1','x2')))

for(i in 1:nrep)
{
set.seed(new_seeds[i])
X <- cbind(1,matrix(rnorm(n*(p-1)),n,p-1))
U <- matrix(rep(rnorm(n,0,sqrt(sigma2u)),w),n,w,byrow=FALSE)
E <- matrix(rnorm(n*w,0,sqrt(sigma2)),n,w)
Y <- matrix(X\%*\%beta,n,w,byrow=FALSE) + U + E

mixdata <- data.frame(y = as.vector(Y),
X = I(kronecker(as.matrix(rep(1,w)),X[,-1])),
id = as.vector(ID))

bmlmer_res <- bmlmer(formula = bmlmerform,data = mixdata,
                      nsim = 5000,nu1=0.01,nu2=0.01,beta0=NULL,nu0=0.01)

lme4_res <- confint(lme4::lmer(formula = bmlmerform,data = mixdata))

eval[1,] <- eval[1,] + ((bmlmer_res$Quantiles[1,] < par_vec)&
                        (bmlmer_res$Quantiles[3,] > par_vec))
                        
eval[2,] <- eval[2,] + ((lme4_res[,1] < par_vec)&
                        (lme4_res[,2] > par_vec))
}

print(eval/nrep)

}



}
\references{
Gangsei, L.E.& Vinje, H. 2025. A closed form solution for
Bayesian analysis of a simple linear mixed model.
}
\author{
Hilde Vinje\cr
hilde.vinje@nmbu.no\cr
+47 959 01 450

Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
